+++ src/main/java/si/matjazcerkvenik/alertmonitor/web/uibeans/UiConfigBean.java
@@ -117,21 +117,6 @@ public class UiConfigBean implements Serializable {
         return configParamsList;
     }
 
-//    @Deprecated
-//    public String getPromServer() {
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getProviderConfig().getParam(PrometheusDataProvider.DP_PARAM_KEY_SERVER);
-//    }
-//
-//    @Deprecated
-//    public void setPromServer(String server) {
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        if (server.endsWith("/")) server = server.substring(0, server.length()-1);
-//        adp.getProviderConfig().setParam(PrometheusDataProvider.DP_PARAM_KEY_SERVER, server);
-//        LogFactory.getLogger().info("UiConfigBean: prometheus server changed: " + server);
-////		Growl.showInfoGrowl("Configuration updated", "");
-//    }
-
     public void setDataRetention(String time) {
         try {
             AmProps.ALERTMONITOR_DATA_RETENTION_DAYS = Integer.parseInt(time);
@@ -171,17 +156,13 @@ public class UiConfigBean implements Serializable {
         return s;
     }
 
-//    public void setHttpReadTimeout(String interval) {
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        adp.getProviderConfig().setParam(PrometheusDataProvider.DP_PARAM_KEY_CLIENT_READ_TIMEOUT_SEC, interval);
-//        LogFactory.getLogger().info("UiConfigBean: http client read timeout changed: " + interval);
-//    }
-//
-//    public String getHttpReadTimeout() {
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getProviderConfig().getParam(PrometheusDataProvider.DP_PARAM_KEY_CLIENT_READ_TIMEOUT_SEC);
-//    }
+    public String getMongoDbConnectTimeout() {
+        return Integer.toString(AmProps.ALERTMONITOR_MONGODB_CONNECT_TIMEOUT_SEC);
+    }
 
+    public String getMongoDbReadTimeout() {
+        return Integer.toString(AmProps.ALERTMONITOR_MONGODB_READ_TIMEOUT_SEC);
+    }
 
     public void setKafkaEnabled(boolean kafkaEnabled) {
         AmProps.ALERTMONITOR_KAFKA_ENABLED = kafkaEnabled;
@@ -211,25 +192,6 @@ public class UiConfigBean implements Serializable {
         return AmProps.ALERTMONITOR_KAFKA_TOPIC;
     }
 
-//    public String reloadPrometheusAction() {
-//
-//        LogFactory.getLogger().debug("UiConfigBean: reloadPrometheusAction called");
-//
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        PrometheusHttpClient api = adp.getHttpClientPool().getClient();
-//
-//        try {
-//            api.reload();
-//        } catch (Exception e) {
-//            LogFactory.getLogger().error("UiConfigBean: reloadPrometheusAction exception: ", e);
-//        } finally {
-//            adp.getHttpClientPool().returnClient(api);
-//        }
-//
-//        return "";
-//
-//    }
-
 
 
 
@@ -237,78 +199,7 @@ public class UiConfigBean implements Serializable {
 
     /* STATISTICS */
 
-//    public long getWhMsgCount() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getWebhookMessagesReceivedCount();
-//    }
 
-//    public long getJournalCount() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getJournalCount();
-//    }
-
-//    public long getJournalSize() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getJournalSize();
-//    }
-//
-//    public long getAlarmsCount() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getRaisingEventCount();
-//    }
-
-//    public long getClearsCount() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getClearingEventCount();
-//    }
-
-//    public void setSyncInterval(String interval) {
-//        // TODO implement
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        Integer i = Integer.parseInt(interval);
-//        adp.getProviderConfig().setParam(PrometheusDataProvider.DP_PARAM_KEY_SYNC_INTERVAL_SEC, String.valueOf(i));
-//        LogFactory.getLogger().info("UiConfigBean: sync interval changed: " + i);
-////		Growl.showInfoGrowl("Configuration updated", "");
-//        adp.restartSyncTimer();
-//    }
-//
-//    public String getSyncInterval() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getProviderConfig().getParam(PrometheusDataProvider.DP_PARAM_KEY_SYNC_INTERVAL_SEC);
-//    }
-
-//    public String getLastSyncTime() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return Formatter.getFormatedTimestamp(adp.getLastSyncTimestamp(), AmDateFormat.TIME); }
-//
-//    public String getSyncSuccessCount() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return Integer.toString(adp.getSyncSuccessCount()); }
-
-//    public String getSyncFailedCount() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return Integer.toString(adp.getSyncFailedCount()); }
-//
-//    public int getActiveAlarmsCount(String severity) {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getActiveAlarmsList(severity).size();
-//    }
-//
-//    public int getAllActiveAlarmsCount() {
-//        // TODO moved to provider
-//        AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
-//        return adp.getActiveAlerts().size();
-//    }
 
     public int getAllAlarmingInstancesCount() {
         AbstractDataProvider adp = DAO.getInstance().getDataProvider(selectedDataProvider);
@@ -367,7 +258,6 @@ public class UiConfigBean implements Serializable {
         } catch (IOException e) {
             System.out.println("IOException " + e.getMessage());
         }
-//        return "/active";
     }
 
 }