+++ src/main/java/si/matjazcerkvenik/alertmonitor/web/uibeans/UiInstanceBean.java
@@ -15,6 +15,8 @@
  */
 package si.matjazcerkvenik.alertmonitor.web.uibeans;
 
+import org.primefaces.model.timeline.TimelineEvent;
+import org.primefaces.model.timeline.TimelineModel;
 import si.matjazcerkvenik.alertmonitor.data.DAO;
 import si.matjazcerkvenik.alertmonitor.model.DEvent;
 import si.matjazcerkvenik.alertmonitor.model.DTarget;
@@ -28,6 +30,8 @@ import javax.faces.bean.ManagedProperty;
 import javax.faces.bean.ViewScoped;
 import javax.faces.context.FacesContext;
 import java.io.Serializable;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -122,6 +126,86 @@ public class UiInstanceBean implements Serializable {
     public String getUpStatusTooltip() {
         if (target.isUp()) return "At least one job is up";
         return "All jobs are down";
+    }
+
+
+
+    private TimelineModel<String, ?> model;
+    private LocalDateTime start;
+    private LocalDateTime end;
+
+    public TimelineModel<String, ?> getModel() {
+
+        List<DEvent> dEvents = getInstanceJournalAlarms();
+
+        HashMap<String, List<DEvent>> tempMap = new HashMap<>();
+
+        for (DEvent e : dEvents) {
+            System.out.println("getModel: event: " + e.toString());
+            List<DEvent> list = tempMap.getOrDefault(e.getCorrelationId(), new ArrayList<>());
+            list.add(e);
+            tempMap.put(e.getCorrelationId(), list);
+        }
+
+        // create timeline model
+        model = new TimelineModel<>();
+
+        // set initial start / end dates for the axis of the timeline
+        start = LocalDate.of(-140, 1, 1).atStartOfDay();
+        end = LocalDate.of(-140, 1, 2).atStartOfDay();
+
+        for (String s : tempMap.keySet()) {
+            for (DEvent e : tempMap.get(s)) {
+                TimelineEvent event = TimelineEvent.builder()
+                        .data(e.getInfo())
+                        .startDate(start)
+                        .endDate(end)
+                        .editable(true)
+                        .group(e.getAlertname())
+                        .styleClass("sdfg".toLowerCase())
+                        .build();
+
+                model.add(event);
+            }
+        }
+
+
+
+//        // groups
+//        String[] names = new String[]{"User 1", "User 2", "User 3", "User 4", "User 5", "User 6"};
+//
+//        for (String name : names) {
+//            LocalDateTime end = start.minusHours(12).withMinute(0).withSecond(0).withNano(0);
+//
+//            for (int i = 0; i < 5; i++) {
+//                LocalDateTime start = end.plusHours(Math.round(Math.random() * 5));
+//                end = start.plusHours(4 + Math.round(Math.random() * 5));
+//
+//                long r = Math.round(Math.random() * 2);
+//                String availability = (r == 0 ? "Unavailable" : (r == 1 ? "Available" : "Maybe"));
+//
+//                // create an event with content, start / end dates, editable flag, group name and custom style class
+//                TimelineEvent event = TimelineEvent.builder()
+//                        .data(availability)
+//                        .startDate(start)
+//                        .endDate(end)
+//                        .editable(true)
+//                        .group(name)
+//                        .styleClass(availability.toLowerCase())
+//                        .build();
+//
+//                model.add(event);
+//            }
+//        }
+        return model;
+    }
+
+    public LocalDateTime getStart() {
+        return start;
+    }
+
+    public LocalDateTime getEnd() {
+        return end;
     }
 
 }